House Cleaning for Engineers and A Player Timer
Added a Timer Widget which is hooked up to the new PlayerController subclass located in Blueprints/Core/.
To Set the time the timer from anywhere simply use "GetPlayerController"  (index 0). Cast to our Controller class and then use the SetPlayerTimer function in seconds.

As for organization:
Blueprints/Core/ contains:
- FirstPersonGameMode (set the new default Player controller to the OurGamesController class)
- FirstPersonHUD
- OurGamesController

Blueprints/UI contains:
- Just PlayerCounterDownTimer UserWidget at the moment.

Hopeful this is clear. Dm me if anything is confusing